Split index update and full-scan runners
